requirements.txt
Flask==3.0.0
Pillow==10.1.0
requests==2.31.0
Werkzeug==3.0.1
python-telegram-bot==20.7
reportlab==4.0.7
